feat(S01/T01): extract state.ts and helpers.ts from webview monolith
- state.ts (123 lines): all types (ToolCallState, TurnSegment, AssistantTurn,
  ChatEntry, AppState, AvailableModel, ToolCategory), shared state object, nextId()
- helpers.ts (277 lines): escapeHtml, formatting, tool helpers, markdown rendering,
  subagent output, scrollToBottom, marked configuration
- Removed dead code: followUpQueue (declared never used), hasStructuredDetails
- scrollToBottom now takes container param instead of closure capture
- index.ts reduced from 2,149 to 1,721 lines
- Build passes, all imports resolve
